requirements.txt
python-telegram-bot==20.8
python-dotenv==1.0.1
aiohttp==3.9.3
pytz==2024.1
emoji==2.10.1
pandas
openpyxl
xlrd 